+++ xenon-gateway/src/main/java/com/vmware/xenon/gateway/GatewayPathFactoryService.java
@@ -35,19 +35,13 @@ public class GatewayPathFactoryService extends FactoryService {
     }
 
     @Override
-    protected String buildDefaultChildSelfLink(ServiceDocument document) {
-        // A specific format for the self-link is used to guarantee
-        // uniqueness of uri paths for a specific gateway.
+    protected String buildDefaultChildSelfLink(ServiceDocument document)
+            throws IllegalArgumentException {
         GatewayPathService.State state = (GatewayPathService.State)document;
-        if (state.path != null) {
-            return createSelfLinkFromState(state);
+        if (state.path == null) {
+            throw new IllegalArgumentException("path is required");
         }
-
-        if (state.documentSelfLink != null) {
-            return state.documentSelfLink;
-        }
-
-        return super.buildDefaultChildSelfLink();
+        return createSelfLinkFromState(state);
     }
 
     public static final String createSelfLinkFromState(GatewayPathService.State state) {